requirements.txt
pure-python-adb
PyYaml
